multiselect.service to util function and replaced it usage appropriately everywhere.
