+++ projects/multiselect/src/lib/grouped-options/grouped-options.component.spec.ts
@@ -1,11 +1,10 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { DebugElement, TemplateRef, ElementRef, Renderer2 } from '@angular/core';
+import { DebugElement, TemplateRef, ElementRef } from '@angular/core';
 
 import { GroupedOptionsComponent } from './grouped-options.component';
 import { NgxMultiselectComponent } from '../multiselect.component';
-import { NgxMultiselectService } from '../services/multiselect.service';
 import { VirtualScrollDirective } from '../directives/virtual-scroll.directive';
 import { CommonModule } from '@angular/common';
 import { GroupByMultiselectOption, MultiselectOption } from '../models/multiselect-option.model';
@@ -45,7 +44,7 @@ describe('Grouped Options Component', () => {
     selectedFirstOption = options.filter(o => o.id);
     component.multiple = true;
     component._selectedOptions = [] as MultiselectOption[];
-    multiselect = new NgxMultiselectComponent(<ElementRef<any>>null, new NgxMultiselectService());
+    multiselect = new NgxMultiselectComponent(<ElementRef<any>>null);
     multiselect.multiple = true;
     multiselect.setOptions(options);
     component.options = multiselect.getOptions();
