+++ projects/multiselect/src/public_api.ts
@@ -5,7 +5,7 @@
 export * from './lib/multiselect.module';
 export * from './lib/multiselect.component';
 export * from './lib/pipes/display-selected-value.pipe';
-export * from './lib/services/multiselect.service';
+export * from './lib/utils.spec';
 export * from './lib/grouped-options/grouped-options.component';
 export * from './lib/options/options.component';
 export * from './lib/filter-options/filter-options.component';
